Added endpoints for creating and logging into an account
Created REST endpoints that allow will allow users to create an account and log into it after. Stores the account information within the data base and when logging in checks to make sure the passwords match. Passwords are encrypted using BCrypt as after some research it seems like one of the most secure ways to encrypt a password.