Sends verification email and allows users to verify account
When the user creates an account it will send them an email with a generated code. They are then able to enter that code to verify their email address. Still have to add in functionality for resending a verification code and checking that a users account is verified before they log in.