Added endpoint to allow user to select a destination
Added the necessary components that allow users to start building a trip. Once user selects a destination it will create a trip within the database linked to their account and to a destination they selected. Updated endpoint documentation and added comment to remind to revisit the account delete function when trips and comments are done.